Improved scheduling information output

+++ include/hal.h
@@ -28,7 +28,7 @@ void cpu_slow(void);
 /**
  * Generate a random number between r_min and r_max.
  */
-uint32_t my_random_in_range(uint32_t r_min, uint32_t r_max);
+int64_t my_random_in_range(uint32_t r_min, uint32_t r_max);
 
 #endif 
 /* EOF */
